+++ calcipy/experiments/bump_programmatically.py
@@ -1,15 +1,14 @@
 """Experiment with bumping the git tag using `griffe`."""
 
-import sys
-
 import griffe
 import semver
 from beartype import beartype
 from corallium.log import logger
+from griffe.exceptions import BuiltinModuleError
 
 
 @beartype
-def bump_tag(*, pkg_name: str, tag: str, tag_prefix: str, debug: bool = True) -> str:
+def bump_tag(*, pkg_name: str, tag: str, tag_prefix: str) -> str:
     """Make a SemVer minor bump using `griffe` if there were any breaking changes.
 
     Major versions must be bumped manually
@@ -19,25 +18,18 @@ def bump_tag(*, pkg_name: str, tag: str, tag_prefix: str, debug: bool = True) ->
     current = griffe.load(pkg_name)
 
     breakages = [*griffe.find_breaking_changes(previous, current)]
+    for breakage in breakages:
+        try:
+            logger.text(breakage._explain_oneline())  # noqa: SLF001
+        except BuiltinModuleError:  # noqa: PERF203
+            logger.warning(str(breakage))
+        except Exception:
+            logger.exception(str(breakage))
 
-    if debug:
-        for breakage in breakages:
-            logger.debug(breakage._explain_oneline())  # noqa: SLF001
-
-    ver = semver.Version.parse(current_tag.replace(tag_prefix, ''))
+    try:
+        ver = semver.Version.parse(tag.replace(tag_prefix, ''))
+    except ValueError:
+        logger.exception('Failed to parse tag', tag=tag)
+        return ''
     new_ver = ver.bump_minor() if any(breakages) else ver.bump_patch()
     return f'{tag_prefix}{new_ver}'
-
-
-if __name__ == '__main__':
-    """Run as a standalone file without calcipy:
-
-    ```sh
-    poetry run python bump_programmatically.py "3.0.7" | tail -n 1
-    ```
-
-    """
-    current_tag = sys.argv[1]
-    new_tag = bump_tag(pkg_name='pytest_cache_assert', tag=current_tag, tag_prefix='v')
-    logger.warning('')
-    logger.warning(new_tag)